+++ go-backend/main.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"joshsoftware/peerly/aws"
 	"joshsoftware/peerly/config"
-	"joshsoftware/peerly/cronjobs"
 	"joshsoftware/peerly/db"
 
 	"joshsoftware/peerly/service"
@@ -94,9 +93,6 @@ func startApp() (err error) {
 
 	// mux router
 	router := service.InitRouter(deps)
-
-	// Start up all the background tasks Peerly depends upon
-	cronjobs.Init(deps)
 	// init web server
 	server := negroni.Classic()
 	server.UseHandler(router)
