+++ go-backend/tasks/tasks.go
@@ -13,6 +13,8 @@ import (
 // external scripts that would otherwise do the same thing.
 func Init(deps service.Dependencies) {
 	s1 := gocron.NewScheduler(time.UTC)
+	s1.Every(1).Day().At("00:00").Do(deps.Store.ResetHi5QuotaBalanceJob)
+
 	s1.Every(1).Hours().Do(deps.Store.CleanBlacklistedTokens)
 	s1.Start()
 }
